Fix re-rendering in content scripts in Firefox extensions
There is a problem with Reagent in content scripts in Firefox extensions where
the UI does not update when a ratom changes.

The problem is described here
https://github.com/NoxHarmonium/reagent/commit/a9e898cd74b53fd3321fe7e122fb3d1ddef91495
with Reagent 0.9.0+, though that Reagent fork also has a branch that patches
Reagent 0.8.1. The relevant branches are:
* fix-raf-binding-0.8.1
* fix-raf-binding-0.9.0-rc1
* master (with fix on 0.9.0)

Note the "related links section" of the commit message at the above link: React
issue https://github.com/facebook/react/issues/16606
From what I’ve read, this only seems to affect content scripts in browser
extensions in Firefox.

I have merely applied the fix in the commit at the above link to Reagent 0.10.0.

[#170650669]